Class and Structures Documentation: 17/09
Added documentation to explain in detail the Object Oriented approach that was implemented to store and perform the exploration of the competition's map. Formatted and changed the directories of files to correctly separate them by the zone they are used in (inside the general directory are the files that are used throughout the entire round).