Debugging and Testing: 20/09
Debugged a lot of code today regarding the Vertex structure and the Graph class. Since I am using hashmaps with Vertex objects as keys to save information about the graph and its traversal, I had to specify a process for the Vertex object to be hashed. After doing so I leveraged Github Copilot to write testing cases on a C++ file because I have limited experience working with unit tests and testcase development. After these tests I am 100% sure that the class and structures are ready to use as they are intended in order to start developing each zone's algorithm