Debugging and Testing Algorithm: 22/09
Going further on the last commit, the originally presented graph and structures presented difficulties when triying to implement the algorithm and any graph structure because of the redundant hashmaps. These were serving the same purpose as just saving more attributes on the vertex structures, and also occupying more memory, as more objects were initialized to store the data. With this improved structure, every necessary attribute is found inside every Vertex object, and not in hashmaps, and the Graph class was changed to struct to facilitate the retrieval and editing of its data. After developing these changes, a Graph_Test.cpp file was created to perform tests on the new methods and on the new implementation. Once the test was successful, a practical test was created using the current testing map structure available at Roborrego's HQ. By creating a simple DFS algorithm adapted to the Vertex and Graph structs, a not-so-optimized path to the end was obtained.