Arduino Port, Robot Struct, and Start Zona A Final Implementation
In this commit I first focused on porting all of the code I had written so that it was compatible with Arduino IDE. In the process some variables, such as std::strings, were changed to fit Arduino's own workings. There is a bug regarding 'new_handler' within the ArduinoSTL Library, used to import the C++ Standard Library to Arduino, and I had to change 'Arduino AVR Boards' inside Arduino IDE to version 1.8.2. Once the code was compiled successfully, I started implementing the actual algorithm that the robot will use to traverse Zone A by sending and receiving data to the robots components. While the specific functions for each component's usage are still pending, developing the algorithm before-hand can allow me to really comprehend the specific functions needed to complete each stage.