Port Code to Arduino: 27/09
Successfully ported the code to Arduino Mega, as this is probably one of the only boards that can support the memory use involved in the algorithm. To do this I changed all std::string to String() (from the Arduino library), improved memory usage by replacing some basic structures to simple arrays, and reduced the dependency on the Standard Library.