Final day: Ramp, Zone B and Zone A recap. 30/09
During the last two days before the competition began I focused on compatibility between the Functions.h header file and the code I had already written. By doing so, I was able to start the final implementation of the Zone A algorithm, which in the end would not work only because of the size of the robot and a malfunction on one of its ultrasonic sensors. The ramp method described on the Ramp.h header file corresponds to a simple implementation to traverse this area without touching the outside walls. The Zone B algorithm was left TODO as our main focus was making sure the robot could correctly move and interact with its surroundings, rather than the specific implementations of these elements.